refactor(scheduling): rename scheduling to task_utils
1. It better describes the actual functionality - these are utilities for working with tasks
2. "Scheduling" implies more about the actual scheduling mechanism, which is handled by Celery
3. These functions are task-related helpers rather than scheduling logic
4. It follows the pattern of other utility modules (like price_display.py, notifications.py)
